+++ src/MockHttp/Matchers/VersionMatcher.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Net.Http;
+
+namespace MockHttp.Matchers
+{
+	/// <summary>
+	/// Matches a request by the HTTP message version.
+	/// </summary>
+	public class VersionMatcher : ValueMatcher<Version>
+	{
+		/// <summary>
+		/// Initializes a new instance of the <see cref="VersionMatcher"/> class using specified <paramref name="version"/>.
+		/// </summary>
+		/// <param name="version"></param>
+		public VersionMatcher(Version version) : base(version ?? throw new ArgumentNullException(nameof(version)))
+		{
+		}
+
+		/// <inheritdoc />
+		public override bool IsMatch(HttpRequestMessage request)
+		{
+			return request.Version.Equals(Value);
+		}
+
+		/// <inheritdoc />
+		public override string ToString()
+		{
+			return $"Version: {Value}";
+		}
+
+		/// <inheritdoc />
+		public override bool IsExclusive => true;
+	}
+}